toolchain: remove scripts
Signed-off-by: Dennis Liu <dennis48161025@gmail.com>

The purpose of `scripts` is to ensure that executables are correctly
placed in `install/<package>/lib`. By default, `setuptools` places the
executables corresponding to `entry_points` in `install/<package>/bin`.
By adding a `setup.cfg` file, you can control the behavior of
`setuptools` and force the executables to be placed in
`install/<package>/lib`, which is also the default location where ROS 2
looks for them.